api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "numpy>=2.3.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
    "prometheus-client>=0.21.1",
]
